fix: display parent span trace if any
