+++ src/client/src/app/(playground)/requests/page.tsx
@@ -1,5 +1,4 @@
 "use client";
-import RequestTable from "./request-table";
 import { useCallback, useEffect } from "react";
 import RequestFilter from "./request-filter";
 import { RequestProvider } from "./request-context";
@@ -9,7 +8,7 @@ import { toast } from "sonner";
 import { getFilterDetails } from "@/selectors/filter";
 import { useRootStore } from "@/store";
 import { getPingStatus } from "@/selectors/database-config";
-import { Separator } from "@/components/ui/separator";
+import List from "./list";
 
 export default function RequestPage() {
 	const filter = useRootStore(getFilterDetails);
@@ -40,7 +39,7 @@ export default function RequestPage() {
 	return (
 		<RequestProvider>
 			<RequestFilter total={(data as any)?.total} />
-			<RequestTable
+			<List
 				data={(data as any)?.records || []}
 				isFetched={isFetched || pingStatus !== "pending"}
 				isLoading={isLoading || pingStatus === "pending"}
