+++ src/client/src/app/api/metrics/request/parent-span/[id]/route.ts
@@ -1,0 +1,13 @@
+import { getRequestViaParentSpanId } from "@/lib/platform/request";
+
+export async function GET(_: Request, context: any) {
+	const { id } = context.params || {};
+
+	if (!id)
+		return Response.json("No parent span id provided", {
+			status: 400,
+		});
+
+	const res: any = await getRequestViaParentSpanId(id);
+	return Response.json(res);
+}
